# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Digit Killer - Deriv Trading Analysis API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: symbols
    description: Active symbols
  - name: analysis
    description: Digit analysis
  - name: signals
    description: AI signals
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /active-symbols:
    get:
      operationId: getActiveSymbols
      tags: [symbols]
      summary: Get all active symbols from Deriv
      responses:
        "200":
          description: List of active symbols
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActiveSymbol"

  /digit-analysis:
    get:
      operationId: getDigitAnalysis
      tags: [analysis]
      summary: Get digit frequency analysis for a symbol
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: count
          in: query
          required: false
          schema:
            type: integer
            default: 1000
      responses:
        "200":
          description: Digit distribution analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DigitAnalysis"

  /wide-eye-analysis:
    get:
      operationId: getWideEyeAnalysis
      tags: [analysis]
      summary: Get wide eye D-circles analysis (1000 ticks)
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: count
          in: query
          required: false
          schema:
            type: integer
            default: 1000
      responses:
        "200":
          description: Wide eye D-circles analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WideEyeAnalysis"

  /over-under-signals:
    get:
      operationId: getOverUnderSignals
      tags: [signals]
      summary: Get over/under entry signals and recommendations
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: count
          in: query
          required: false
          schema:
            type: integer
            default: 1000
      responses:
        "200":
          description: Over/under analysis with entry points
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverUnderSignals"

  /even-odd-analysis:
    get:
      operationId: getEvenOddAnalysis
      tags: [analysis]
      summary: Get even/odd digit analysis
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: count
          in: query
          required: false
          schema:
            type: integer
            default: 1000
      responses:
        "200":
          description: Even/odd analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvenOddAnalysis"

  /match-differ-signals:
    get:
      operationId: getMatchDifferSignals
      tags: [signals]
      summary: Get best match/differ digit recommendations
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: count
          in: query
          required: false
          schema:
            type: integer
            default: 1000
      responses:
        "200":
          description: Match/differ signals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchDifferSignals"

  /tick-contracts:
    get:
      operationId: getTickContracts
      tags: [analysis]
      summary: Get tick contracts analysis (rise/fall, high/low tick, only up/down)
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: count
          in: query
          required: false
          schema:
            type: integer
            default: 1000
      responses:
        "200":
          description: Tick contracts analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TickContractsAnalysis"

  /ai-signals:
    get:
      operationId: getAiSignals
      tags: [signals]
      summary: Get AI-generated trading signals
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: AI signals with full analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSignals"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials

  /admin/users:
    get:
      operationId: getUsers
      tags: [admin]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppUser"
    post:
      operationId: createUser
      tags: [admin]
      summary: Create a new user (generate user ID)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"

  /admin/users/{id}:
    delete:
      operationId: deleteUser
      tags: [admin]
      summary: Revoke/delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: User deleted

  /admin/users/{id}/revoke:
    patch:
      operationId: revokeUser
      tags: [admin]
      summary: Revoke user access
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"

  /user/login:
    post:
      operationId: userLogin
      tags: [admin]
      summary: User login with generated ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCredentials"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSession"
        "401":
          description: Invalid credentials

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ActiveSymbol:
      type: object
      required: [symbol, display_name, market, submarket, is_open, pip_size]
      properties:
        symbol:
          type: string
        display_name:
          type: string
        market:
          type: string
        submarket:
          type: string
        is_open:
          type: integer
        pip_size:
          type: number

    DigitFrequency:
      type: object
      required: [digit, count, percentage, rank]
      properties:
        digit:
          type: integer
        count:
          type: integer
        percentage:
          type: number
        rank:
          type: integer
        color:
          type: string

    DigitAnalysis:
      type: object
      required: [symbol, count, digits, current_digit, most_frequent, least_frequent]
      properties:
        symbol:
          type: string
        count:
          type: integer
        digits:
          type: array
          items:
            $ref: "#/components/schemas/DigitFrequency"
        current_digit:
          type: integer
        most_frequent:
          type: integer
        least_frequent:
          type: integer
        current_price:
          type: number
        even_count:
          type: integer
        odd_count:
          type: integer
        even_pct:
          type: number
        odd_pct:
          type: number

    DCircle:
      type: object
      required: [symbol, count, digits, current_digit, green_arc, blue_arc, red_arc, yellow_arc]
      properties:
        symbol:
          type: string
        count:
          type: integer
        digits:
          type: array
          items:
            $ref: "#/components/schemas/DigitFrequency"
        current_digit:
          type: integer
        green_arc:
          type: integer
        blue_arc:
          type: integer
        red_arc:
          type: integer
        yellow_arc:
          type: integer
        current_price:
          type: number

    WideEyeAnalysis:
      type: object
      required: [symbol, d_circle_1000, d_circle_custom, over_under, even_odd]
      properties:
        symbol:
          type: string
        d_circle_1000:
          $ref: "#/components/schemas/DCircle"
        d_circle_custom:
          $ref: "#/components/schemas/DCircle"
        custom_count:
          type: integer
        over_under:
          $ref: "#/components/schemas/OverUnderSummary"
        even_odd:
          $ref: "#/components/schemas/EvenOddSummary"
        current_price:
          type: number

    OverUnderSummary:
      type: object
      required: [threshold, under_count, over_count, under_pct, over_pct, equal_count, equal_pct]
      properties:
        threshold:
          type: integer
        under_count:
          type: integer
        over_count:
          type: integer
        equal_count:
          type: integer
        under_pct:
          type: number
        over_pct:
          type: number
        equal_pct:
          type: number

    EvenOddSummary:
      type: object
      required: [even_count, odd_count, even_pct, odd_pct]
      properties:
        even_count:
          type: integer
        odd_count:
          type: integer
        even_pct:
          type: number
        odd_pct:
          type: number
        current_digit:
          type: integer

    OverUnderEntry:
      type: object
      required: [contract, recommended_ticks, risk_level, entry_digit, confidence, conditions_met]
      properties:
        contract:
          type: string
        recommended_ticks:
          type: string
        risk_level:
          type: string
        entry_digit:
          type: integer
        confidence:
          type: number
        conditions_met:
          type: boolean
        reason:
          type: string

    OverUnderSignals:
      type: object
      required: [symbol, digits, best_over, best_under, entries, current_price, current_digit]
      properties:
        symbol:
          type: string
        digits:
          type: array
          items:
            $ref: "#/components/schemas/DigitFrequency"
        best_over:
          $ref: "#/components/schemas/OverUnderEntry"
        best_under:
          $ref: "#/components/schemas/OverUnderEntry"
        entries:
          type: array
          items:
            $ref: "#/components/schemas/OverUnderEntry"
        current_price:
          type: number
        current_digit:
          type: integer

    EvenOddAnalysis:
      type: object
      required: [symbol, even_count, odd_count, even_pct, odd_pct, current_digit, recommended, ticks, confidence]
      properties:
        symbol:
          type: string
        even_count:
          type: integer
        odd_count:
          type: integer
        even_pct:
          type: number
        odd_pct:
          type: number
        current_digit:
          type: integer
        recommended:
          type: string
        ticks:
          type: integer
        confidence:
          type: number
        recent_digits:
          type: array
          items:
            type: integer

    MatchDifferSignals:
      type: object
      required: [symbol, best_match, best_differ, match_ticks, differ_ticks, digits, current_digit]
      properties:
        symbol:
          type: string
        best_match:
          type: integer
        best_differ:
          type: integer
        match_ticks:
          type: integer
        differ_ticks:
          type: integer
        match_confidence:
          type: number
        differ_confidence:
          type: number
        digits:
          type: array
          items:
            $ref: "#/components/schemas/DigitFrequency"
        current_digit:
          type: integer
        reason_match:
          type: string
        reason_differ:
          type: string

    TickContractEntry:
      type: object
      required: [contract_type, recommended_ticks, confidence, entry_signal, description]
      properties:
        contract_type:
          type: string
        recommended_ticks:
          type: integer
        confidence:
          type: number
        entry_signal:
          type: string
        description:
          type: string
        risk_level:
          type: string

    TickContractsAnalysis:
      type: object
      required: [symbol, current_price, current_digit, contracts, trend, volatility_score]
      properties:
        symbol:
          type: string
        current_price:
          type: number
        current_digit:
          type: integer
        contracts:
          type: array
          items:
            $ref: "#/components/schemas/TickContractEntry"
        trend:
          type: string
        volatility_score:
          type: number
        recent_ticks:
          type: array
          items:
            type: number

    AiSignal:
      type: object
      required: [id, symbol, contract_type, direction, entry_digit, ticks, confidence, strategy, timestamp]
      properties:
        id:
          type: string
        symbol:
          type: string
        contract_type:
          type: string
        direction:
          type: string
        entry_digit:
          type: integer
        ticks:
          type: integer
        confidence:
          type: number
        strategy:
          type: string
        timestamp:
          type: string
        reason:
          type: string
        risk_level:
          type: string

    AiSignals:
      type: object
      required: [symbol, signals, last_updated]
      properties:
        symbol:
          type: string
        signals:
          type: array
          items:
            $ref: "#/components/schemas/AiSignal"
        last_updated:
          type: string
        market_condition:
          type: string

    AdminCredentials:
      type: object
      required: [password]
      properties:
        password:
          type: string

    AdminSession:
      type: object
      required: [token, role]
      properties:
        token:
          type: string
        role:
          type: string

    UserCredentials:
      type: object
      required: [user_id, password]
      properties:
        user_id:
          type: string
        password:
          type: string

    UserSession:
      type: object
      required: [token, user_id, username]
      properties:
        token:
          type: string
        user_id:
          type: string
        username:
          type: string

    UserInput:
      type: object
      required: [username]
      properties:
        username:
          type: string

    AppUser:
      type: object
      required: [id, user_id, username, active, created_at]
      properties:
        id:
          type: integer
        user_id:
          type: string
        username:
          type: string
        active:
          type: boolean
        created_at:
          type: string
        revoked_at:
          type: ["string", "null"]