# lib/api-zod/src/generated/types/aiSignal.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Digit Killer - Deriv Trading Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface AiSignal {
  id: string;
  symbol: string;
  contract_type: string;
  direction: string;
  entry_digit: number;
  ticks: number;
  confidence: number;
  strategy: string;
  timestamp: string;
  reason?: string;
  risk_level?: string;
}