# lib/api-zod/src/generated/types/appUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Digit Killer - Deriv Trading Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface AppUser {
  id: number;
  user_id: string;
  username: string;
  active: boolean;
  created_at: string;
  /** @nullable */
  revoked_at?: string | null;
}