# lib/api-zod/src/generated/types/digitAnalysis.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Digit Killer - Deriv Trading Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { DigitFrequency } from './digitFrequency';

export interface DigitAnalysis {
  symbol: string;
  count: number;
  digits: DigitFrequency[];
  current_digit: number;
  most_frequent: number;
  least_frequent: number;
  current_price?: number;
  even_count?: number;
  odd_count?: number;
  even_pct?: number;
  odd_pct?: number;
}