# lib/api-zod/src/generated/types/overUnderSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Digit Killer - Deriv Trading Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface OverUnderSummary {
  threshold: number;
  under_count: number;
  over_count: number;
  equal_count: number;
  under_pct: number;
  over_pct: number;
  equal_pct: number;
}