# lib/api-zod/src/generated/types/userCredentials.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Digit Killer - Deriv Trading Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface UserCredentials {
  user_id: string;
  password: string;
}